I added in downsampling
and again the transformer was only giving one result.  It seems
like it was a problem with the normalization of the time input.
I expanded the precision to double and that seems to have fixed it
but you never know...

+++ app/src/main/java/com/example/runeswipe/model/Models.kt
@@ -269,7 +269,7 @@ class Player(
 data class Point(
     val x: Float,
     val y: Float,
-    val t: Float   // time in milliseconds or relative time
+    val t: Double   // time in milliseconds or relative time
 )
